new components for user page
Previous bugs related to the http post request of the vehicle, haven't been solved yet.

In the meanwhile, user page has been implemented with the addition of components for booking a vehicle and the one for taking a look on active and previous bookings.
Few things still need to be made here too.